Fixed Xmask
Xmask was looking for consecutive months/years. This was a problem for the independent inflection method, because they would split and the program would skip over everything because it was just reading 0s. This was fixed by combining all trend columns and always taking the max value.
Inclusion of inflection index list for future multi-inflection points.
Changing the Comparison of proxy data and ozone data, so that proxy data will be cut to the exact time frame.