Trend slope and coefficient overhaul
Streamlining of trend line in program plotting. Depending on the string that corresponds to each coefficient, either a mean is calculated (month-of-the-year) or the first coefficient will be used.
The same was done for the calculation of the trend value.
There still seems to be a problem that the trend line is too high.