Date as Year + Month
Added the option to input two columns/variables for the date. First is the year and second is the month. This works for both netCDF (year_var, month_var) or ascii (year_col, month_col).
The UI supports this already but there is not yet a way to input the two date variables/columns.