Data Limits Feature and Alternative Contour Plotting
Added a feature to the config files as well as the user interface to chose the limits of each dimension. This will help when only looking at trends in specific regions without needing to compute trends for all regions.
The limits will also be used for the plotting routine, so that each figure will only show that specific region.
Added an alternative contour plot.