2D-Proxy Fix and Toolbar for Plots
Previously when taking an averaging window for a 2d proxy, the program would take the mean over the chosen months but also over all latitudes of that proxy if it was 2d. This error was removed and it will now only take the mean over the months and not any additional dimension if the proxy has those.

Added Toolbars to each plot, to easier change plots on the fly and also zoom and move to see specific features of the plot.

The proxy contour plot is not yet updated to include inflection point plots.